Backend: Refactor rate controller to make it "thinner"
Add a simple "glue" layer between the adapters and the core (kind of a
dependency injector) to allow the rate service to obtain the appropriate
rate-backend implementation without having to receive it as an argument.
In this way service clients have to do less work in order to call the
service.